refactor(core): extract enumerator + technique_registry, wire assess to service layer
Split technique_engine.py (463 lines) into three modules:
  technique_registry.py (127) — TechniqueDefinition + TechniqueRegistry
  technique_engine.py   (206) — execute_technique + probe execution
  enumerator.py         (141) — enumerate_targets for MCP tool

Wire CLI assess command to kimera.core.assessor instead of the old
SecurityScanner. Assessment now uses config-driven checks from
config/checks/workload.yaml with Rich table rendering.

Convert C2 (secret enumeration) and C3 (configmap extraction) from
exec-mode shell scripts to API-mode techniques — these test RBAC
permissions which the K8s API can check directly without exec.

Expand _enumerate_for_technique to handle all resource types from the
technique registry: secrets, configmaps, namespaces, networkpolicies,
rolebindings, clusterrolebindings, endpoints.

All core modules now under 300 lines. 387 tests, zero lint errors.
