+++ config/techniques/C2-secret-enumeration.yaml
@@ -1,7 +1,6 @@
 # Technique: Secret enumeration via API
 # ID: C2
 # Tactic: Credential Access
-# Uses the pod's SA token to list secrets via the K8s API.
 
 id: C2
 name: Secret enumeration via API
@@ -14,53 +13,29 @@ mitre:
   data_sources: [Container, User Account]
 
 cis_controls:
-  - "5.1.2"  # Minimize access to secrets
-  - "5.1.3"  # Minimize wildcard use in Roles
+  - "5.1.2"
+  - "5.1.3"
 
 severity: critical
 
 description: |
-  Uses the pod's mounted ServiceAccount token to query the Kubernetes API
-  for secrets in the namespace. Reports the count and names of accessible
-  secrets. Database credentials, API keys, and TLS certificates are commonly
+  Tests whether the configured kubeconfig credentials can list secrets in
+  the target namespace. Reports count and names of accessible secrets.
+  Database credentials, API keys, and TLS certificates are commonly
   exposed through this vector.
 
+  This technique uses the K8s API directly (not exec), testing the same
+  permissions an attacker with stolen kubeconfig or SA token would have.
+
 execution:
-  mode: exec
-  requires_target_pod: true
+  mode: api
 
-  probes:
-    - type: command
-      run: |
-        SA_DIR="/var/run/secrets/kubernetes.io/serviceaccount"
-        if [ ! -f "$SA_DIR/token" ]; then echo "NO_TOKEN"; exit 0; fi
-        TOKEN=$(cat "$SA_DIR/token")
-        NS=$(cat "$SA_DIR/namespace")
-        API="https://kubernetes.default.svc.cluster.local"
-        RESP=$(wget -q -O- --header="Authorization: Bearer $TOKEN" \
-            --no-check-certificate "$API/api/v1/namespaces/$NS/secrets" 2>&1)
-        if echo "$RESP" | grep -q '"items"'; then
-            COUNT=$(echo "$RESP" | grep -c '"name"' || true)
-            echo "SECRET_COUNT:$COUNT"
-            echo "$RESP" | grep '"name"' | head -10 | \
-                sed 's/.*"name": "\(.*\)".*/SECRET_NAME:\1/'
-        else
-            echo "ACCESS_DENIED"
-        fi
-
-  evidence_markers:
-    - marker: "SECRET_COUNT"
-      evidence: "SA can enumerate secrets in namespace"
-      impact: "Database credentials, API keys, and TLS certificates exposed"
-    - marker: "SECRET_NAME"
-      evidence: "Individual secret name discovered"
-    - marker: "ACCESS_DENIED"
-      evidence: "API server denied secrets access to this SA"
-    - marker: "NO_TOKEN"
-      evidence: "No SA token mounted in pod"
+  api_calls:
+    - resource: secrets
+      verb: list
 
 success_indicators:
-  - "SECRET_COUNT"
+  - "secrets_found"
 
 impact:
   - "Database credentials, API keys, and TLS certificates exposed"
